#164 Move ShaftEscape.OBSERVABLE to GameObservable
Because of this, it resembles the Singleton pattern. Plus, not half of
all classes have dependencies on the ShaftEscape application anymore.

<organization,naming
