+++ src/main/java/nl/tudelft/ti2206/group9/audio/SoundEffectPlayer.java
@@ -1,6 +1,6 @@
 package nl.tudelft.ti2206.group9.audio;
 
-import static nl.tudelft.ti2206.group9.ShaftEscape.OBSERVABLE;
+import static nl.tudelft.ti2206.group9.util.GameObservable.OBSERVABLE;
 
 import java.io.File;
 import java.net.MalformedURLException;
