+++ src/main/java/nl/tudelft/ti2206/group9/level/CollisionHandler.java
@@ -6,7 +6,7 @@ import nl.tudelft.ti2206.group9.level.entity.AbstractPickup;
 import nl.tudelft.ti2206.group9.level.entity.Player;
 import nl.tudelft.ti2206.group9.level.entity.PowerupInvulnerable;
 import nl.tudelft.ti2206.group9.util.GameObserver;
-import static nl.tudelft.ti2206.group9.ShaftEscape.OBSERVABLE;
+import static nl.tudelft.ti2206.group9.util.GameObservable.OBSERVABLE;
 
 /**
  * Add all the collision handlers to a CollisionMap.
