+++ src/main/java/nl/tudelft/ti2206/group9/level/State.java
@@ -1,6 +1,6 @@
 package nl.tudelft.ti2206.group9.level;
 
-import static nl.tudelft.ti2206.group9.ShaftEscape.OBSERVABLE;
+import static nl.tudelft.ti2206.group9.util.GameObservable.OBSERVABLE;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import nl.tudelft.ti2206.group9.gui.skin.Skin;
